One massive commit
Just doing this hurts my soul.  But I had no idea how anything in the template
was structured, and kept needing to jump from thing to thing while gaining an
understanding of its structure.  By the time that understanding was gained about
half the project was already done.  At that point it was simply not worth the
effort to separate the work into multiple commits.